Set up first spec with autoload of files
We created a generic board class, I honestly don't know if this is
actually where I want to start, because either way this will be
complicated and I need to diagram it up.
But now the tests are running and it's "working", and the files are all
autoloaded inside of rspec. Maybe that's dumb but it works for Rails.

- Add guard for continuous test running.
- Add turnip for acceptance tests for something new.