Add abilitiy to set players for a game
This lays foundation for a game actually being able to start.
We have the generic idea of being able to set players, with a
faction type into a game.

There's now an option for different kinds of boards but I don't think
that'll ever get used, but I don't consider it over engineering, as long
as I don't go down the rabbit hole of abstracting any code out of Boards
into base.

- PlayerList contains all players and is a facade to handling getting
  victory points or knowing who the current player is and etc.
- I added basic skeletons for factions, but they do nothing atm.