Add concept of a Deck, with supporting Card object
The game now sets up and loads a deck.
The deck right now is all the same card, and does not do anything.

The cards also have no real quantity except for the suit, again,
skeleton framework for the time being on cards. Next step is going to be
each players faction getting their tokens setup on the board.