Add sandi_meter and fix code to mostly follow it
I want to try a couple of neewww things on this one.
First thing, going to follow the sandi rules, that's the neat one to see
how small I can get all of this object and classes and how I can make
the API nice.

I ALSO want to get Sorbet added on to this soon, because really how
often do I get to work on a PURE ruby project? Getting to play with
types would be fun.