Start bird setup
Birds now have their little meeps and roosts!
I've handled placement of it for both where there is and is not a cat.

Next steps is going to be getting a decree handled, and also adding a
leader, that's def a 2 part combo because the leader directly affects
where the viziers go.

- Changed meeple symbol to 'o', because 6 'm' side by side looked too
  crowded. Might need to play with this but it's ok for now imo.
- Added more methods to handle finding a corner with a certain token.
  Really only needed for the keep and initial roost, but that's ok.