Rework game setup to take different keyword args
Woo this is a doozy. Okay, so now whenever game calls setup, we take in
different sets of keyword arguments. This has a pretty big code change
in specs, but overally not much actually changes because before we were
doing (x = nil) and other garbage like that anyway.

We do this for specific faction setup, some need a shared set of cards,
some need just the quest deck, some need the player list, it's all
crazy.

Passing in the entire Game though means we'd need to initialize
ERRYTHING else, just to check how something interacts with the board, or
with a quest deck. Not worth it.

- This also introduces the concept of a quest deck
- This removes basic deck functionality out into a Base class, so that
  more things can use it as needed.