Add ability to craft cards to cats
Okay! So this handles the very first step of daylight, the ability to
craft. To do that I needed to add Items to the board, checking for if
the item is available, and if I have enough craftable clearings in my
hand. I also need to make sure that someone can't use a suit twice, so
if they use a mouse for one card, they can't use it again.

- Added a discard to decks
- Renamed rabbit to bunny because I kept using both, but bunny is right