Start doing options and moves for a cats turn
We now have a list of available options, context sensitive and correct.
We can't do anything with them, nor are they plugged in yet :sweats: but
this is a good foundation for handling things like this.

I might need to do a major refactor that passes in the board and deck at
all times, on player create or etc. It doesn't really make sense, if I
need them EVERY time I just need them. Most things only make sense in
terms of how they effect the board.

- Minor change in the terminal inputs, we raise and rescue a specific
  error. Otherwise we catch the input not valid in guard error
  repeatedly and break.