Start reworking options to have 2 different paths
First we check if we CAN do something, separatly from if the act of
actually doing it. This way we can test either part away from each
other, which is super helpful. Then we can just test that it does the
thing correctly, entirely separate from if we should stop due to
irrelevant conditions.