Add battling and removing tokens
Okay this was another big ol sweaty one.
I'm at the point where I NEED to refactor.

But this should be an entire faction done, at least with minimum base
faction board rules. Obviously need to display things like VP and Cards
and w.e. relevant info to make it playable, but at least everything is
in there.

I have at least 3 big refactors to work on next, so it's time to figure
out how that's going to work and what that should look like.