Add ability to use birds in decree
Bird cards count as wild cards in the decree now, and can be used.
This also required the user to pick which card they wanted to use, when
having the option to use a bird card or a non bird card.

This may be reworkable actually, could just always assume they want to
use bird card last as it has the most flexibility. Hmm.