Move Battle logic to Faction Base, add to Decree
The battle logic had actually already been mostly handled by the cats,
so to transfer it all that was needed to move it into faction base. The
logic for actually deciding where to battle is slightly, different since
one is resolved in the decree and one just picks, but the _BATTLE_ logic
is the same and reusable.