Summary of Async and Curses experiment
Async is just hard. It's possibly doable, but it's just not worth the
time and effort, for a game that comes out soon enough.
Curses was the more disappointing one, it doesn't handle colors very
well, so, screw it, whatever.

The experimenting did result in something very helpful, which is
properly figuring out cursor escape sequences. We now are able to use
those, and remove screen clear, which removes screen clear ENTIRELY.

At this point, I think the best way to get this in a usable state is to
get this with the CWE bots as AI, and every faction playable.