Starting leaders for birds
With this and the Field Hospital done, the Cats and Birds will be
entirely done. That's 2 factions that are _COMPLETELY_ done.

There's still other game functionality, but at this point I'd like to
take a break and continue on to the Mice faction. Once that's done, the
Vagabondo will help force Items to be a necessary addition, so that's
pretty neat / terrifying.