Add mobilize and train to Mouse brah
Nice, okay this is simple and it's back to 100/100.
No real issues with any of this.

The only 'annoying' thing is actually the difference between craft
items, and the other 2 options.
Craft items puts you into the craft 'mode', while the others are
singular options. I'd like to move them all into modes. It feels nicer.