The Commit where RSpec DSL is added
It's been done, this ain't thoughtbot style guide anymore. I'm okay with
it though, I feel like the reasoning for why is valid on this:

That being it became extremely cumbersome to change tests, or really
write new ones. The tests are still pretty obvious what is happening,
'faction' is the current faction we're testing.
bird_player and etc are also obvious, and everything is starting to
consolidate and work the same way. Hopefully this means any big
refactors in the code will align with the tests.