Big one, better error, cats done, new token icon
Error now just reappears at the beginning of the prompt, but doesn't
push the options down.

I think a valid solution is just "this is a big game make your font size
smaller" for now. There's obviously major accessibility concerns there,
but I'm okay rocking with it for now as it's a solo project only lol.

Also, tbh, if it goes onto a website that'll be _so_ much easier to make
work and not deal with this.

Cats have all their major options done for being able to go back! Lots
of small things also need changes, but that's after the big change
refactor where all of this branch coverage just goes into that one
method.

meeples are already 'o's, and all tokens are now lowercase to separate from
buildings. Maybe there's another, better way, but maybe not.