Start adding favor card
Going to take a pause and do a refactor very similar to the
deck / board refactor for the players list. It's needed in too many
places now, because so many things need to account for the other
players. The other possible refactor for this is that all pieces know
who their owner is, so they can just be sent "remove" and it's handled.