Refactored player to work similar to board / deck
Slightly more annoying than board and deck, but that's ok for now.
This allows for much easier item actions and all of that, since player
no longer needs to be passed around. To interact with someone you need
to have knowledge that they exist, so I believe this is ok.