Removing allied meeple does not make hostile
This is to resolve a bug where being the defender in a battle did not
make you hostile. This is wrong, you're still hostile with them if you
remove a unit.

The thing is, if you removed an allied warrior, there's a different
condition for that. The later code block covers that.