Add Improvments and Armorers
Quick side note, the reason for the return if no choice is effectively
just to handle the ambush. I'm sure it can be expanded to handle more
places, but that is not for the current discussion so whatever.

Beyond that, added improvements! Minor bug, you should not be able to
craft multiple of one type of improvement, so that still needs work.
Improvements also do not get exhuasted, because the only one currently
available gets discarded anyway, not a big deal.