Add CommandWarren
This is a little tricky, and not ideal.
This needs to be refactored, to allow for cancelling. Right now, if
crafted, and used, and cancelled, it just stops. That's not the ideal.

Flow should be:
pick improvement to use
do improvement
exhaust improvement

cancel would go back to picking improvement to use