Add ability to craft 'any' crafts
Until now the crafts were to be a specific suit, and would handle that.
But Royal Claim, and maybe more in the E&P deck don't work that way.
So currently this only handles royal_claim, and could be reworked in the
future if necessary to plug in more cards with an 'any'. But for now,
not too important.