I've hit a deal breaking point and idk if fixable
This example is removing wood.
Either the game state needs to change, for example removing 2 wood, but
it can not change, because of type null problems.

For example, you need to remove 2 wood, an much earlier yield might
remove them both. This is the issue with "yielding" everything.

Either everything is changeable and things break, or nothing is
changeable and deeply nested transactions don't work.

I also have problems with the issue of handling things that need to be
chained, for example march, or anything related to outrage and movement.

Putting this online seems incredibly hard, and possibly not worth the
effort?